64-bit transition part 13: the rest of ksh93
This is the thirteenth of the thickfold patch series, which enables ksh93
to operate within a 64-bit address space.

This patch applies the rest of the changes to ksh93
necessary for 64-bit memory allocation, along with
the usual warning fixes.

Notable changes:
- parse.c: Fixed a build error when using SHOPT_KIA (re: 63366e34).
- sh_pipe(): Fix broken SOCK_CLOEXEC check. The !SOCK_CLOEXEC check
  in the #if directive apparently didn't work, as strace was showing
  unnecessary fcntl calls were still getting compiled in. Check for
  SOCK_CLOEXEC==0 in a non-preprocessor if statement and let the
  compiler optimize it away, which actually works correctly.
- Also removed some unnecessary ptrdiff_t casts.
- Add a (Linux-specific) regression test for allocating 5GB (we check
  if plenty of RAM is available via /proc/meminfo).
- locale.sh: Disable a test that requires SHOPT_MULTIBYTE to work
  correctly.

Change in the number of warnings on Linux when compiling with clang using
-Wsign-compare -Wshorten-64-to-32 -Wsign-conversion -Wimplicit-int-conversion:
352 => 37 (progression from part 12 => part 13)

The nvflags warnings that remain will be addressed in a separate
follow-up patch of their own (which doesn't affect memory allocation).

Fixes https://github.com/ksh93/ksh/issues/592

+++ src/cmd/ksh93/include/path.h
@@ -33,12 +33,12 @@
 #   endif
 #endif /* !SHOPT_SPAWN */
 
-#define PATH_PATH		0001
-#define PATH_FPATH		0002
-#define PATH_CDPATH		0004
-#define PATH_BFPATH		0010
-#define PATH_SKIP		0020
-#define PATH_STD_DIR		0100	/* directory is on  $(getconf PATH) */
+#define PATH_PATH		0001U
+#define PATH_FPATH		0002U
+#define PATH_CDPATH		0004U
+#define PATH_BFPATH		0010U
+#define PATH_SKIP		0020U
+#define PATH_STD_DIR		0100U	/* directory is on $(getconf PATH) */
 
 #define PATH_OFFSET	2		/* path offset for path_join */
 #define MAXDEPTH	1024		/* maximum shell function recursion depth */
@@ -57,8 +57,8 @@ typedef struct pathcomp
 	char		*lib;
 	char		*bbuf;
 	char		*blib;
-	unsigned short	len;
-	unsigned short	flags;
+	size_t		len;
+	uint16_t	flags;
 } Pathcomp_t;
 
 #ifndef _ARGNOD_H
@@ -69,7 +69,7 @@ typedef struct pathcomp
 extern void		path_newdir(Pathcomp_t*);
 extern Pathcomp_t	*path_dirfind(Pathcomp_t*,const char*,int);
 extern Pathcomp_t	*path_unsetfpath(void);
-extern Pathcomp_t	*path_addpath(Pathcomp_t*,const char*,int);
+extern Pathcomp_t	*path_addpath(Pathcomp_t*,const char*,uint16_t);
 extern Pathcomp_t	*path_dup(Pathcomp_t*);
 extern void		path_delete(Pathcomp_t*);
 extern void 		path_settrackedalias(const char*,Pathcomp_t*);
@@ -77,7 +77,7 @@ extern Namval_t		*path_gettrackedalias(const char*);
 extern Pathcomp_t 	*path_absolute(const char*, Pathcomp_t*, int);
 extern char 		*path_basename(const char*);
 extern char 		*path_fullname(const char*);
-extern int 		path_expand(const char*, struct argnod**, int);
+extern size_t 		path_expand(const char*, struct argnod**, int);
 extern noreturn void 	path_exec(const char*,char*[],struct argnod*);
 extern pid_t		path_spawn(const char*,char*[],char*[],Pathcomp_t*,int);
 extern int		path_open(const char*,Pathcomp_t*);
@@ -86,9 +86,9 @@ extern char 		*path_pwd(void);
 extern Pathcomp_t	*path_nextcomp(Pathcomp_t*,const char*,Pathcomp_t*);
 extern int		path_search(const char*,Pathcomp_t**,int);
 extern char		*path_relative(const char*);
-extern int		path_complete(const char*, const char*,struct argnod**);
+extern size_t		path_complete(const char*, const char*,struct argnod**);
 #if SHOPT_BRACEPAT
-    extern int 		path_generate(struct argnod*,struct argnod**, int);
+    extern ssize_t	path_generate(struct argnod*,struct argnod**, int);
 #endif /* SHOPT_BRACEPAT */
 
 #if SHOPT_DYNAMIC
